Spike for moving Purchase contract creation out into a library.
This means that every Listing no longer has to
contain the entire Purchase contract binary code. 
This makes a big cut in the cost to create a new listing.

This spike is not intended for use as is. It here so 
we can have a look at a simple version of the idea.